pyproject.toml
[tool.poetry]
name = "tinyagent"
version = "0.1.0"
description = ""
authors = ["Konstantin.Grotov <konstantin.grotov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymupdf = "1.23.26"
bs4 = "0.0.1"
fastapi = "0.110.1"
langchain-community = "0.0.9"
langchain-core = "0.1.7"
langchain-openai = "0.0.2"
langchain = "0.1.0"
numexpr = "2.8.7"
openai = "1.6.1"
protobuf = "5.26.1"
python-dateutil = "2.8.2"
sentencepiece = "0.2.0"
tiktoken = "0.5.2"
torch = "2.2.2"
transformers = "^4.44.2"
uvicorn = "0.29.0"
python-multipart = "0.0.9"
httpx = "0.27.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
